sg: extract ci command in its own package
