+++ dev/sg/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
         "sg_analytics.go",
         "sg_app.go",
         "sg_audit.go",
-        "sg_ci.go",
         "sg_cloud.go",
         "sg_db.go",
         "sg_deploy.go",
@@ -49,9 +48,9 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/dev/sg",
     visibility = ["//visibility:private"],
     deps = [
-        "//dev/ci/runtype",
         "//dev/sg/adr",
         "//dev/sg/buf",
+        "//dev/sg/ci",
         "//dev/sg/dependencies",
         "//dev/sg/internal/analytics",
         "//dev/sg/internal/background",
@@ -64,7 +63,6 @@ go_library(
         "//dev/sg/internal/generate/golang",
         "//dev/sg/internal/generate/proto",
         "//dev/sg/internal/images",
-        "//dev/sg/internal/loki",
         "//dev/sg/internal/migration",
         "//dev/sg/internal/open",
         "//dev/sg/internal/repo",
@@ -109,11 +107,9 @@ go_library(
         "@com_github_agext_levenshtein//:levenshtein",
         "@com_github_buildkite_go_buildkite_v3//buildkite",
         "@com_github_docker_docker_credential_helpers//credentials",
-        "@com_github_gen2brain_beeep//:beeep",
         "@com_github_gomodule_redigo//redis",
         "@com_github_google_go_github//github",
         "@com_github_google_go_github_v41//github",
-        "@com_github_google_uuid//:uuid",
         "@com_github_grafana_regexp//:regexp",
         "@com_github_jackc_pgx_v4//:pgx",
         "@com_github_keegancsmith_sqlf//:sqlf",
