+++ dev/sg/main.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/sourcegraph/log"
 
+	"github.com/sourcegraph/sourcegraph/dev/sg/ci"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/analytics"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/background"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/secrets"
@@ -272,7 +273,7 @@ var sg = &cli.App{
 		// Common dev tasks
 		startCommand,
 		runCommand,
-		ciCommand,
+		ci.Command,
 		testCommand,
 		lintCommand,
 		generateCommand,
